% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{chHeatmap}
\alias{chHeatmap}
\title{Hmisc heatmap wrapper function}
\usage{
chHeatmap(
  mat,
  cell_mm = 3,
  col_and_row_name_size = 8,
  .col = circlize::colorRamp2(c(-1, -0.5, 0, 0.5, 1), c("#444EB8", "#A2A7DB", "white",
    "#D38A90", "#B8444E")),
  .name = "r",
  heatmap_legend_param = NULL,
  ...
)
}
\description{
Hmisc heatmap wrapper function
}
