% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{clusterSizeAnno}
\alias{clusterSizeAnno}
\title{annotate cluster size on complexheatmap}
\usage{
clusterSizeAnno(
  x,
  col = "bisque3",
  .text_size = 6,
  .width = 0.3,
  .name_size = 6
)
}
\description{
annotate cluster size on complexheatmap
}
