% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{enricher_wrapper}
\alias{enricher_wrapper}
\title{Over-representation analysis}
\usage{
enricher_wrapper(target_genes, bkg_genes, geneset_df, geneset_info)
}
\arguments{
\item{bkg_genes}{character vector of all genes expressed in the dataset}

\item{geneset_df}{dataframe form of geneset IDs and genes. e.g., gs_ls$genesets_and_genes}

\item{geneset_info}{dataframe with info such as names, descriptions, database, etc. gs_ls$geneset_info}
}
\description{
Over-representation analysis
}
