% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{fgsea_wrapper}
\alias{fgsea_wrapper}
\title{Wrapper function for fgsea ranked geneset enrichment analysis}
\usage{
fgsea_wrapper(input_genes, geneset_ls, geneset_info, scoreType = "pos")
}
\description{
Wrapper function for fgsea ranked geneset enrichment analysis
}
