% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{pdfEmbed}
\alias{pdfEmbed}
\title{Export figures in PDF files and then embed fonts}
\usage{
pdfEmbed(
  file,
  plot_fxn,
  width,
  height,
  family = "ArialMT",
  device = "quartz",
  bg = "white",
  ...
)
}
\arguments{
\item{file}{Destination filename. Must be a full path, not the ~/ if referring to a different file location from the working directory.}

\item{plot_fxn}{A function that prints the plot. E.g., \code{plot(saved_ggplot)}}

\item{width}{Width of file to save, in unit inches.}

\item{height}{Height of file to save, in unit inches.}
}
\description{
A note about pdf saving issues - if there are kerning or sizing issues, use this function so that plots are saved with PDFs then the font is embedded in the file
This function takes the steps to save the plot as a pdf with the arial font, and then embedding the font in the file. this should not change the sizing of fonts and should return normal kerning. it seems like with any PDF, the font might look a little more "bold" than when saved as a png
}
\examples{
# example code
}
\keyword{internal}
