% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns.R
\name{hmSigDots}
\alias{hmSigDots}
\title{Add dots to heatmap}
\usage{
hmSigDots(value_mat, threshold = 0.05, dot_pt_size = 2, fun = "layer")
}
\arguments{
\item{value_mat}{provide a significance matrix the same dimensions and order as the matrix used to plot values}

\item{threshold}{a significance score to determine what cells to print a dot}

\item{dot_pt_size}{size of dot to print on heatmap}

\item{fun}{value of either layer or cell}
}
\description{
Annotate significant ComplexHeatmap Heatmap cells with dot if significant. This can be used as a layer function in the \code{Heatmap} function.
}
